Undo split tests code into 2 classes.
